Pk ifix department entity service (#32)
* IFIX-311: Create Department Hierarchy

* Merge branch 'ifix-department-entity-service' of https://github.com/egovernments/iFix-Dev into pk-ifix-department-entity-service

# Conflicts:
#	domain-services/ifix-department-entity-service/src/main/resources/application.properties

* IFIX-313: Department entity create api

* changed the level data tpe to Integer

* IFIX-314: Search Department Hierarchy

+++ domain-services/ifix-department-entity-service/src/main/java/org/egov/web/models/DepartmentHierarchyLevelSearchCriteria.java
@@ -37,7 +37,7 @@ public class DepartmentHierarchyLevelSearchCriteria {
     private String label = null;
 
     @JsonProperty("level")
-    private BigDecimal level = null;
+    private Integer level = null;
 
 
     public DepartmentHierarchyLevelSearchCriteria addIdsItem(String idsItem) {
